fix: decouple live comment types from sessions
